split backend_gtk3 into two parts
_backend_gtk3.py does not depend on `Gcf`.

backend_gkt3.py does depend on `Gcf` and maintains reverse compatibility
